dashboard page structure, nav button padding
