+++ apps/searchneu/components/navigation/NavBar.tsx
@@ -48,7 +48,7 @@ export function NavBar({
         <Link
           href="/scheduler"
           data-active={pathname === "/scheduler"}
-          className="bg-neu1 data-[active=true]:border-neu3 flex w-full items-center rounded-full border-1 p-2 text-sm"
+          className="bg-neu1 data-[active=true]:border-neu3 flex w-full items-center rounded-full border-1 px-4 p-2 text-sm"
         >
           Scheduler
         </Link>
@@ -57,7 +57,7 @@ export function NavBar({
         <Link
           href="/graduate"
           data-active={pathname === "/graduate"}
-          className="bg-neu1 data-[active=true]:border-neu3 flex w-full items-center rounded-full border-1 p-2 text-sm"
+          className="bg-neu1 data-[active=true]:border-neu3 flex w-full items-center rounded-full border-1 px-4 p-2 text-sm"
         >
           Graduate
         </Link>
